Fix serialization after .debug()
