+++ src/index.ts
@@ -1,3 +1,9 @@
+import type { Modifier } from './modifiers';
+import type { Operation } from './operations';
+import type { Shape } from './shapes/index';
+import type { Transformation } from './transformations/index';
+
+export type Scad = Modifier | Operation | Shape | Transformation;
+
 export * from './operations';
-export * from './ScadCommand';
 export * from './shapes/index';
