+++ test/rounded_objects.test.ts
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import { rounded_cube, rounded_square } from '../src';
+import { rounded_cube, rounded_square } from '../src/index';
 
 const test_rounded_square = (size = 1, r = 0.125) => ({
   type: 'translate',
